Add function to extract decays from pdg-ascii-file
Start to implement the function to extract decays from the
pdg-ascii-file and write them systematically in a new file. It now works
for D*(2010)+- but there is still a lot of work to be done.
Also changed the function name extract_decays to
extract_decays_from_decays
This may break the tests, I will fix it in a later commit
also, open_file_safely() should be moved to a utility.py module