Add function to extract decays from line and tests
The reference file will have to be parsed one decay at a time. To be
better able to test and maintain this process, use a separate function
to turn one string possibly spanning multiple lines in the reference
file into a tuple of list of daughters, branching fraction and branching
fraction errors. This development can only be decently done via test
driven development. Also add dependency to numpy.

The next step will be to place another function before this one, which
turns a multi-line string into a one-line string, which should be easier
to parse.