Remove prints and add string post-processing
Added some post-processing to remove unnecessary whitespace and check
that there is only one item per list and then return this item as
string, since there is no point in returning a one-item list.
Also I have decided to return all 4 columns, since this is more stable
to test and easier to work with later on.