Skip lines starting with '['
I am not yet sure if this is correct behaviour, but currently I cannot
find these decays in the decay.dec. Since these decays are skipped, I
added the tests, but commented them out. The maybe necessary if these
decays are relevant and have to be added again to the code.