Add second translation for anti-K*0
Apparently the pdg ascii file uses K*0 and K*(892)0 interchangeably. The
same is true for Kbar*0 and Kbar*(892)0. However now we have two
translations for anti-K*0. I am not sure what will, or what should
happen if the translation from EvtGen to pdg for anti-K*0 is attempted.
However it seems that K*(892)0 is the deminant variant.