Make flag-remove a func., add it to 1st col. creat
As can be seen in the added test, sometimes there is a flag which does
not intruduce a line-break but the decay daughters in front of it do.
Hence the flags must also be removed during reassembling the first
column. To not duplicate code, the removel has been moved to its own
function.