Replace K_L0 with K_S0 if decay not found
The Decay.dec contains many cases where a decay with both K_S0 and K_L0
is presented, but only the K_S0 decay is listed by the pdg. Since the
values for both should be roughly equal, the value for the decay with
K_S0 is used if the Decay with K_L0 is not found. However it is unclear
where the original K_L0 value comes from.